+++ signaldata.cpp
@@ -22,18 +22,18 @@
 
 using namespace std;
 
-SignalData::SignalData(uint64_t samplerate) :
+SignalData::SignalData(double samplerate) :
 	_samplerate(samplerate),
 	_start_time(0)
 {
 }
 
-uint64_t SignalData::get_samplerate() const
+double SignalData::get_samplerate() const
 {
 	return _samplerate;
 }
 
-int64_t SignalData::get_start_time() const
+double SignalData::get_start_time() const
 {
 	return _start_time;
 }
